+++ packages/cms/lib/modules/resource-form-widgets/lib/fields.js
@@ -196,6 +196,11 @@ const fields = [
                       label: "User: neighborhood",
                       showFields: ['fieldRequired']
                   },
+                  {
+                      value: 'userIsPublic',
+                      label: "User: public status",
+                      showFields: ['fieldRequired']
+                  },
                 ]
             },
             {