+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 # Changelog
 
 ## v0.9.0 (Pre release)
-
 * Update NPM modules for security
 * Don't add JWT when doing a resource overview call since results are cached
 * Empty cache on DELETE call via api proxy
@@ -22,6 +21,8 @@
 * Remove button-blue class from vote counter
 * Fix wrong error in budgeting
 * Fix previous/next buttons CSS in budgeting
+* Possibility to display top links dependent on user logged in status
+* Allow public profile status (viewablebyRole) to be edited by the user
 
 ## Unreleased updates
 * Fix: Ideas Map had an option to not link Flags. The option had disappeared but is now back.